requirements.txt
Django>=4.2
psycopg2
Pillow
djangorestframework-simplejwt
djangorestframework
celery[redis]